Update ec2_elb_lb.py
ec2_elb_lb check_mode

+++ lib/ansible/modules/cloud/amazon/ec2_elb_lb.py
@@ -716,19 +716,27 @@ class ElbManager(object):
     @_throttleable_operation(_THROTTLING_RETRIES)
     def _delete_elb(self):
         # True if succeeds, exception raised if not
-        result = self.elb_conn.delete_load_balancer(name=self.name)
+        # this will skip deletion in check mode, but doesn't check if a change needs to happen - self.changed will always be true
+        result = True
+        if not module.check_mode:
+            result = self.elb_conn.delete_load_balancer(name=self.name)
         if result:
             self.changed = True
             self.status = 'deleted'
 
     def _create_elb(self):
         listeners = [self._listener_as_tuple(l) for l in self.listeners]
-        self.elb = self.elb_conn.create_load_balancer(name=self.name,
-                                                      zones=self.zones,
-                                                      security_groups=self.security_group_ids,
-                                                      complex_listeners=listeners,
-                                                      subnets=self.subnets,
-                                                      scheme=self.scheme)
+        # this will skip creation in check mode, but doesn't check if a change needs to happen - self.changed will always be true
+        if not module.check_mode:
+            self.elb = self.elb_conn.create_load_balancer(name=self.name,
+                                                          zones=self.zones,
+                                                          security_groups=self.security_group_ids,
+                                                          complex_listeners=listeners,
+                                                          subnets=self.subnets,
+                                                          scheme=self.scheme)
+        else:
+            self.changed = True
+            self.staus = 'created'
         if self.elb:
             # HACK: Work around a boto bug in which the listeners attribute is
             # always set to the listeners argument to create_load_balancer, and
@@ -745,17 +753,23 @@ class ElbManager(object):
 
     def _create_elb_listeners(self, listeners):
         """Takes a list of listener tuples and creates them"""
-        # True if succeeds, exception raised if not
-        self.changed = self.elb_conn.create_load_balancer_listeners(self.name,
-                                                                    complex_listeners=listeners)
+        if not module.check_mode:
+            # True if succeeds, exception raised if not
+            self.changed = self.elb_conn.create_load_balancer_listeners(self.name,
+                                                                        complex_listeners=listeners)
+        # will always be True, no checks are done to determine whether the above call is required
+        self.changed = True
 
     def _delete_elb_listeners(self, listeners):
         """Takes a list of listener tuples and deletes them from the elb"""
         ports = [l[0] for l in listeners]
 
-        # True if succeeds, exception raised if not
-        self.changed = self.elb_conn.delete_load_balancer_listeners(self.name,
-                                                                    ports)
+        if not module.check_mode:
+            # True if succeeds, exception raised if not
+            self.changed = self.elb_conn.delete_load_balancer_listeners(self.name,
+                                                                        ports)
+        # will always be True, no checks are done to determine whether the above call is required
+        self.changed = True
 
     def _set_elb_listeners(self):
         """
@@ -809,10 +823,14 @@ class ElbManager(object):
                 listeners_to_remove.append(existing_listener_tuple)
 
         if listeners_to_remove:
-            self._delete_elb_listeners(listeners_to_remove)
+            if not module.check_mode:
+                self._delete_elb_listeners(listeners_to_remove)
+            self.changed = True
 
         if listeners_to_add:
-            self._create_elb_listeners(listeners_to_add)
+            if not module.check_mode:
+                self._create_elb_listeners(listeners_to_add)
+            self.changed = True
 
     def _api_listener_as_tuple(self, listener):
         """Adds ssl_certificate_id to ELB API tuple if present"""
@@ -845,26 +863,30 @@ class ElbManager(object):
         return tuple(listener_list)
 
     def _enable_zones(self, zones):
-        try:
-            self.elb.enable_zones(zones)
-        except boto.exception.BotoServerError as e:
-            self.module.fail_json(msg='unable to enable zones: %s' % e.message, exception=traceback.format_exc())
+        if not module.check_mode:
+            try:
+                self.elb.enable_zones(zones)
+            except boto.exception.BotoServerError as e:
+                self.module.fail_json(msg='unable to enable zones: %s' % e.message, exception=traceback.format_exc())
 
         self.changed = True
 
     def _disable_zones(self, zones):
-        try:
-            self.elb.disable_zones(zones)
-        except boto.exception.BotoServerError as e:
-            self.module.fail_json(msg='unable to disable zones: %s' % e.message, exception=traceback.format_exc())
+        if not module.check_mode:
+            try:
+                self.elb.disable_zones(zones)
+            except boto.exception.BotoServerError as e:
+                self.module.fail_json(msg='unable to disable zones: %s' % e.message, exception=traceback.format_exc())
         self.changed = True
 
     def _attach_subnets(self, subnets):
-        self.elb_conn.attach_lb_to_subnets(self.name, subnets)
+        if not module.check_mode:
+            self.elb_conn.attach_lb_to_subnets(self.name, subnets)
         self.changed = True
 
     def _detach_subnets(self, subnets):
-        self.elb_conn.detach_lb_from_subnets(self.name, subnets)
+        if not module.check_mode:
+            self.elb_conn.detach_lb_from_subnets(self.name, subnets)
         self.changed = True
 
     def _set_subnets(self):
@@ -911,7 +933,8 @@ class ElbManager(object):
 
     def _set_security_groups(self):
         if self.security_group_ids is not None and set(self.elb.security_groups) != set(self.security_group_ids):
-            self.elb_conn.apply_security_groups_to_lb(self.name, self.security_group_ids)
+            if not module.check_mode:
+                self.elb_conn.apply_security_groups_to_lb(self.name, self.security_group_ids)
             self.changed = True
 
     def _set_health_check(self):
@@ -941,7 +964,8 @@ class ElbManager(object):
                     update_health_check = True
 
             if update_health_check:
-                self.elb.configure_health_check(self.elb.health_check)
+                if not module.check_mode:
+                    self.elb.configure_health_check(self.elb.health_check)
                 self.changed = True
 
     def _check_attribute_support(self, attr):
@@ -957,8 +981,9 @@ class ElbManager(object):
             if attributes.cross_zone_load_balancing.enabled:
                 self.changed = True
             attributes.cross_zone_load_balancing.enabled = False
-        self.elb_conn.modify_lb_attribute(self.name, 'CrossZoneLoadBalancing',
-                                          attributes.cross_zone_load_balancing.enabled)
+        if not module.check_mode:
+            self.elb_conn.modify_lb_attribute(self.name, 'CrossZoneLoadBalancing',
+                                              attributes.cross_zone_load_balancing.enabled)
 
     def _set_access_log(self):
         attributes = self.elb.get_attributes()
@@ -979,12 +1004,14 @@ class ElbManager(object):
                     setattr(attributes.access_log, attr, desired_value)
                     update_access_logs_config = True
             if update_access_logs_config:
-                self.elb_conn.modify_lb_attribute(self.name, 'AccessLog', attributes.access_log)
+                if not module.check_mode:
+                    self.elb_conn.modify_lb_attribute(self.name, 'AccessLog', attributes.access_log)
                 self.changed = True
         elif attributes.access_log.enabled:
             attributes.access_log.enabled = False
             self.changed = True
-            self.elb_conn.modify_lb_attribute(self.name, 'AccessLog', attributes.access_log)
+            if not module.check_mode:
+                self.elb_conn.modify_lb_attribute(self.name, 'AccessLog', attributes.access_log)
 
     def _set_connection_draining_timeout(self):
         attributes = self.elb.get_attributes()
@@ -994,12 +1021,14 @@ class ElbManager(object):
                 self.changed = True
             attributes.connection_draining.enabled = True
             attributes.connection_draining.timeout = self.connection_draining_timeout
-            self.elb_conn.modify_lb_attribute(self.name, 'ConnectionDraining', attributes.connection_draining)
+            if not module.check_mode:
+                self.elb_conn.modify_lb_attribute(self.name, 'ConnectionDraining', attributes.connection_draining)
         else:
             if attributes.connection_draining.enabled:
                 self.changed = True
             attributes.connection_draining.enabled = False
-            self.elb_conn.modify_lb_attribute(self.name, 'ConnectionDraining', attributes.connection_draining)
+            if not module.check_mode:
+                self.elb_conn.modify_lb_attribute(self.name, 'ConnectionDraining', attributes.connection_draining)
 
     def _set_idle_timeout(self):
         attributes = self.elb.get_attributes()
@@ -1007,7 +1036,8 @@ class ElbManager(object):
             if attributes.connecting_settings.idle_timeout != self.idle_timeout:
                 self.changed = True
             attributes.connecting_settings.idle_timeout = self.idle_timeout
-            self.elb_conn.modify_lb_attribute(self.name, 'ConnectingSettings', attributes.connecting_settings)
+            if not module.check_mode:
+                self.elb_conn.modify_lb_attribute(self.name, 'ConnectingSettings', attributes.connecting_settings)
 
     def _policy_name(self, policy_type):
         return __file__.split('/')[-1].split('.')[0].replace('_', '-')  + '-' + policy_type
@@ -1031,15 +1061,18 @@ class ElbManager(object):
         for p in getattr(elb_info.policies, policy_attrs['attr']):
             if str(p.__dict__['policy_name']) == str(policy[0]):
                 if str(p.__dict__[policy_attrs['dict_key']]) != str(policy_attrs['param_value'] or 0):
-                    self._set_listener_policy(listeners_dict)
-                    self._update_policy(policy_attrs['param_value'], policy_attrs['method'], policy_attrs['attr'], policy[0])
+                    if not module.check_mode:
+                        self._set_listener_policy(listeners_dict)
+                        self._update_policy(policy_attrs['param_value'], policy_attrs['method'], policy_attrs['attr'], policy[0])
                     self.changed = True
                 break
         else:
-            self._create_policy(policy_attrs['param_value'], policy_attrs['method'], policy[0])
+            if not module.check_mode:
+                self._create_policy(policy_attrs['param_value'], policy_attrs['method'], policy[0])
             self.changed = True
 
-        self._set_listener_policy(listeners_dict, policy)
+        if not module.check_mode:
+            self._set_listener_policy(listeners_dict, policy)
 
     def select_stickiness_policy(self):
         if self.stickiness:
@@ -1073,15 +1106,17 @@ class ElbManager(object):
                     }
                     policy.append(self._policy_name(policy_attrs['type']))
 
-                    self._set_stickiness_policy(elb_info, listeners_dict, policy, **policy_attrs)
+                    if not module.check_mode:
+                        self._set_stickiness_policy(elb_info, listeners_dict, policy, **policy_attrs)
                 elif not self.module.boolean(self.stickiness['enabled']):
                     if len(elb_info.policies.lb_cookie_stickiness_policies):
                         if elb_info.policies.lb_cookie_stickiness_policies[0].policy_name == self._policy_name(policy_type):
                             self.changed = True
                     else:
                         self.changed = False
-                    self._set_listener_policy(listeners_dict)
-                    self._delete_policy(self.elb.name, self._policy_name(policy_type))
+                    if not module.check_mode:
+                        self._set_listener_policy(listeners_dict)
+                        self._delete_policy(self.elb.name, self._policy_name(policy_type))
 
             elif self.stickiness['type'] == 'application':
                 policy = []
@@ -1099,16 +1134,19 @@ class ElbManager(object):
                         'param_value': self.stickiness['cookie']
                     }
                     policy.append(self._policy_name(policy_attrs['type']))
-                    self._set_stickiness_policy(elb_info, listeners_dict, policy, **policy_attrs)
+                    if not module.check_mode:
+                        self._set_stickiness_policy(elb_info, listeners_dict, policy, **policy_attrs)
                 elif not self.module.boolean(self.stickiness['enabled']):
                     if len(elb_info.policies.app_cookie_stickiness_policies):
                         if elb_info.policies.app_cookie_stickiness_policies[0].policy_name == self._policy_name(policy_type):
                             self.changed = True
-                    self._set_listener_policy(listeners_dict)
-                    self._delete_policy(self.elb.name, self._policy_name(policy_type))
+                    if not module.check_mode:
+                        self._set_listener_policy(listeners_dict)
+                        self._delete_policy(self.elb.name, self._policy_name(policy_type))
 
             else:
-                self._set_listener_policy(listeners_dict)
+                if not module.check_mode:
+                    self._set_listener_policy(listeners_dict)
 
     def _get_backend_policies(self):
         """Get a list of backend policies"""
@@ -1143,11 +1181,13 @@ class ElbManager(object):
 
         # enable or disable proxy protocol
         if ensure_proxy_protocol:
-            self._set_proxy_protocol_policy()
+            if not module.check_mode:
+                self._set_proxy_protocol_policy()
 
         # Make the backend policies so
         for item in replace:
-            self.elb_conn.set_lb_policies_of_backend_server(self.elb.name, item['port'], item['policies'])
+            if not module.check_mode:
+                self.elb_conn.set_lb_policies_of_backend_server(self.elb.name, item['port'], item['policies'])
             self.changed = True
 
     def _get_proxy_protocol_policy(self):
@@ -1164,9 +1204,10 @@ class ElbManager(object):
         proxy_policy = self._get_proxy_protocol_policy()
 
         if proxy_policy is None:
-            self.elb_conn.create_lb_policy(
-                self.elb.name, 'ProxyProtocol-policy', 'ProxyProtocolPolicyType', {'ProxyProtocol': True}
-            )
+            if not module.check_mode:
+                self.elb_conn.create_lb_policy(
+                    self.elb.name, 'ProxyProtocol-policy', 'ProxyProtocolPolicyType', {'ProxyProtocol': True}
+                )
             self.changed = True
 
         # TODO: remove proxy protocol policy if not needed anymore? There is no side effect to leaving it there
@@ -1193,13 +1234,15 @@ class ElbManager(object):
 
         add_instances = self._diff_list(assert_instances, has_instances)
         if add_instances:
-            self.elb_conn.register_instances(self.elb.name, add_instances)
+            if not module.check_mode:
+                self.elb_conn.register_instances(self.elb.name, add_instances)
             self.changed = True
 
         if self.purge_instance_ids:
             remove_instances = self._diff_list(has_instances, assert_instances)
             if remove_instances:
-                self.elb_conn.deregister_instances(self.elb.name, remove_instances)
+                if not module.check_mode:
+                    self.elb_conn.deregister_instances(self.elb.name, remove_instances)
                 self.changed = True
 
     def _set_tags(self):
@@ -1225,7 +1268,8 @@ class ElbManager(object):
                 params['Tags.member.%d.Key' % (i + 1)] = key
                 params['Tags.member.%d.Value' % (i + 1)] = dictact[key]
 
-            self.elb_conn.make_request('AddTags', params)
+            if not module.check_mode:
+                self.elb_conn.make_request('AddTags', params)
             self.changed=True
 
         # Remove extra tags
@@ -1234,7 +1278,8 @@ class ElbManager(object):
             for i, key in enumerate(dictact):
                 params['Tags.member.%d.Key' % (i + 1)] = key
 
-            self.elb_conn.make_request('RemoveTags', params)
+            if not module.check_mode:
+                self.elb_conn.make_request('RemoveTags', params)
             self.changed=True
 
     def _get_health_check_target(self):
@@ -1278,6 +1323,7 @@ def main():
 
     module = AnsibleModule(
         argument_spec=argument_spec,
+        supports_check_mode=True,
         mutually_exclusive = [['security_group_ids', 'security_group_names']]
     )
